refactor(sched): split workspace and context below the server
S2 commit 4/5, one atomic step (the halves share state):

- server/state/workspace.* -> sched/workspace.*: the Workspace is
  artifact state through and through. ReadonlyMode moves to
  server/state/session.h (serving vocabulary) and the parsed value now
  lives on Compiler, written once by the master at initialization.
  CrashBudget (the shared-artifact key ledger) leaves quarantine.h for
  sched/crash_budget.h - contract 6 splits the three crash ledgers
  across layers, and the artifact-key budget belongs to sched.
- server/compiler/context_resolver.* -> sched/context.* (lower half,
  merged with context_cache.h and CommandSource): resolution, synthesis
  and the persisted context-domain state. Session* parameters become a
  ContextUse::{Editor,Background} enum - the resolver only ever used
  pointer presence; append_suffix_include and validate_saved_context
  now take the path_id.
- New server/service/context_service.*: the clice/queryContext,
  currentContext and switchContext handlers, drop_orphaned_choices and
  indicates_missing_context - everything that speaks LSP types or
  touches sessions - driving the resolver through its public surface.
- uri_to_path -> server/transport/uri.*.

Behavior-neutral by test: all four suites green.
