+++ src/sched/context.cpp
@@ -1,4 +1,4 @@
-#include "server/compiler/context_resolver.h"
+#include "sched/context.h"
 
 #include <algorithm>
 #include <chrono>
@@ -9,13 +9,11 @@
 
 #include "command/argument_parser.h"
 #include "command/search_config.h"
-#include "server/state/session_store.h"
 #include "support/filesystem.h"
 #include "support/logging.h"
 #include "syntax/include_resolver.h"
 #include "syntax/preamble_synthesis.h"
 
-#include "kota/ipc/lsp/uri.h"
 #include "llvm/ADT/StringExtras.h"
 #include "llvm/ADT/StringSet.h"
 #include "llvm/Support/FileSystem.h"
@@ -24,27 +22,6 @@
 #include "llvm/Support/xxhash.h"
 
 namespace clice {
-
-namespace lsp = kota::ipc::lsp;
-
-bool indicates_missing_context(llvm::ArrayRef<protocol::Diagnostic> diagnostics) {
-    constexpr static llvm::StringRef codes[] = {
-        "err_unknown_typename",
-        "err_undeclared_var_use",
-        "err_undeclared_var_use_suggest",
-        "err_pp_unterminated_conditional",
-    };
-    for(auto& diag: diagnostics) {
-        if(diag.severity != protocol::DiagnosticSeverity::Error || !diag.code.has_value()) {
-            continue;
-        }
-        auto* code = std::get_if<std::string>(&*diag.code);
-        if(code && llvm::is_contained(codes, *code)) {
-            return true;
-        }
-    }
-    return false;
-}
 
 /// Per-file command selection decision log: which tiers were tried, which one
 /// was hit, and a hash of the final command for correlating later failures.
@@ -89,25 +66,6 @@ static CompileCommand& pick_host_command(Workspace& workspace,
         }
     }
     return results.front();
-}
-
-/// Human-readable summary of the distinguishing flags of a command.
-static std::string flags_label(const CompileCommand& cmd) {
-    auto argv = cmd.to_argv();
-    std::string desc;
-    for(std::size_t j = 0; j < argv.size(); ++j) {
-        llvm::StringRef a(argv[j]);
-        if(a.starts_with("-D") || a.starts_with("-O") || a.starts_with("-std=") ||
-           a.starts_with("-g")) {
-            if(!desc.empty())
-                desc += ' ';
-            desc += argv[j];
-            if((a == "-D" || a == "-O") && j + 1 < argv.size()) {
-                desc += argv[++j];
-            }
-        }
-    }
-    return desc;
 }
 
 HeaderMode ContextResolver::header_mode(llvm::StringRef path, std::uint32_t path_id) const {
@@ -222,7 +180,7 @@ bool ContextResolver::fill_header_context_args(llvm::StringRef path,
                                                std::uint32_t path_id,
                                                std::string& directory,
                                                std::vector<std::string>& arguments,
-                                               Session* session,
+                                               ContextUse use,
                                                std::uint32_t* host_path_id) {
     // Opening one of our own synthesized files (prefix/suffix/snapshot):
     // it is a fragment of the host TU it was synthesized for, so compile
@@ -260,7 +218,7 @@ bool ContextResolver::fill_header_context_args(llvm::StringRef path,
     // diagnostics indicate missing includer state. An explicitly chosen
     // occurrence — even #0 — only has meaning under includer-context
     // semantics, so it forces synthesis regardless of the verdict.
-    const SavedContext* choice = active_choice(session, path_id);
+    const SavedContext* choice = active_choice(use, path_id);
     bool has_host_choice = choice && choice->host_path_id != no_path_id;
     bool synthesize = header_mode(path, path_id) == HeaderMode::NeedsContext ||
                       (has_host_choice && choice->occurrence.has_value());
@@ -272,7 +230,7 @@ bool ContextResolver::fill_header_context_args(llvm::StringRef path,
     // content, so it must be rebuilt). Only editor-facing compiles consult
     // the cache; background indexing must stay independent of per-editor
     // context state, so it resolves fresh every time.
-    if(session) {
+    if(use == ContextUse::Editor) {
         if(auto* cached = header_context(path_id)) {
             bool override_mismatch =
                 has_host_choice && (cached->host_path_id != choice->host_path_id ||
@@ -287,14 +245,14 @@ bool ContextResolver::fill_header_context_args(llvm::StringRef path,
     }
 
     std::optional<HeaderContext> local_ctx;
-    const HeaderContext* ctx_ptr = session ? header_context(path_id) : nullptr;
+    const HeaderContext* ctx_ptr = use == ContextUse::Editor ? header_context(path_id) : nullptr;
     if(!ctx_ptr) {
-        auto resolved = resolve_header_context(path_id, session, synthesize);
+        auto resolved = resolve_header_context(path_id, use, synthesize);
         if(!resolved) {
             LOG_WARN("No CDB entry and no header context for {}", path);
             return false;
         }
-        if(session) {
+        if(use == ContextUse::Editor) {
             ctx_ptr = &(header_contexts[path_id] = std::move(*resolved));
         } else {
             // Background indexing stays independent of per-editor context
@@ -350,7 +308,7 @@ bool ContextResolver::fill_header_context_args(llvm::StringRef path,
 CommandSource ContextResolver::resolve_command(llvm::StringRef path,
                                                std::string& directory,
                                                std::vector<std::string>& arguments,
-                                               Session* session,
+                                               ContextUse use,
                                                std::uint32_t* host_path_id) {
     auto path_id = workspace.path_pool.intern(path);
     llvm::SmallVector<llvm::StringRef, 3> tried;
@@ -365,7 +323,7 @@ CommandSource ContextResolver::resolve_command(llvm::StringRef path,
         auto* cmd = &results.front();
         // Multi-config projects: honor the user's chosen CDB entry, matched
         // by canonical command hash so the choice survives CDB reordering.
-        const SavedContext* choice = active_choice(session, path_id);
+        const SavedContext* choice = active_choice(use, path_id);
         if(choice && choice->host_path_id == no_path_id && !choice->command_hash.empty()) {
             for(auto& candidate: results) {
                 if(canonical_command_hash(candidate.to_string_argv(),
@@ -380,14 +338,14 @@ CommandSource ContextResolver::resolve_command(llvm::StringRef path,
         arguments = cmd->to_string_argv();
     };
 
-    const SavedContext* choice = active_choice(session, path_id);
+    const SavedContext* choice = active_choice(use, path_id);
     bool has_host_choice = choice && choice->host_path_id != no_path_id;
 
     // 1. If the file has an active header context via switchContext, use the
     //    host source's CDB entry with file path replaced and preamble injected.
     if(has_host_choice) {
         tried.push_back("switch_context");
-        if(fill_header_context_args(path, path_id, directory, arguments, session, host_path_id)) {
+        if(fill_header_context_args(path, path_id, directory, arguments, use, host_path_id)) {
             log_command_decision(path, tried, CommandSource::IncludeGraph, arguments);
             return CommandSource::IncludeGraph;
         }
@@ -405,7 +363,7 @@ CommandSource ContextResolver::resolve_command(llvm::StringRef path,
     // 3. No CDB entry — try automatic header context resolution.
     if(!has_host_choice) {
         tried.push_back("include_graph");
-        if(fill_header_context_args(path, path_id, directory, arguments, session, host_path_id)) {
+        if(fill_header_context_args(path, path_id, directory, arguments, use, host_path_id)) {
             log_command_decision(path, tried, CommandSource::IncludeGraph, arguments);
             return CommandSource::IncludeGraph;
         }
@@ -420,8 +378,8 @@ CommandSource ContextResolver::resolve_command(llvm::StringRef path,
     return CommandSource::Fallback;
 }
 
-void ContextResolver::append_suffix_include(const Session& session, std::string& text) {
-    auto* context = header_context(session.path_id);
+void ContextResolver::append_suffix_include(std::uint32_t path_id, std::string& text) {
+    auto* context = header_context(path_id);
     if(!context || context->suffix_path.empty()) {
         return;
     }
@@ -441,7 +399,7 @@ void ContextResolver::append_suffix_include(const Session& session, std::string&
 }
 
 std::optional<HeaderContext> ContextResolver::resolve_header_context(std::uint32_t header_path_id,
-                                                                     Session* session,
+                                                                     ContextUse use,
                                                                      bool synthesize) {
     // Find source files that transitively include this header.
     auto hosts = workspace.dep_graph.find_host_sources(header_path_id);
@@ -455,7 +413,7 @@ std::optional<HeaderContext> ContextResolver::resolve_header_context(std::uint32
     std::uint32_t host_path_id = 0;
     std::optional<std::uint32_t> occurrence;
     std::vector<std::uint32_t> chain;
-    const SavedContext* choice = active_choice(session, header_path_id);
+    const SavedContext* choice = active_choice(use, header_path_id);
     bool has_host_choice = choice && choice->host_path_id != no_path_id;
     if(has_host_choice) {
         auto preferred = choice->host_path_id;
@@ -719,8 +677,7 @@ bool ContextResolver::entry_has_hash(llvm::StringRef entry_path, llvm::StringRef
     return false;
 }
 
-void ContextResolver::validate_saved_context(Session& session) {
-    auto path_id = session.path_id;
+void ContextResolver::validate_saved_context(std::uint32_t path_id) {
     auto path = workspace.path_pool.resolve(path_id);
 
     // A context choice persisted from an earlier session stays authoritative
@@ -745,282 +702,6 @@ void ContextResolver::validate_saved_context(Session& session) {
             saved_contexts.erase(it);
         }
     }
-}
-
-bool ContextResolver::drop_orphaned_choices(SessionStore& sessions) {
-    bool dropped_saved = false;
-    for(auto& [session_id, session]: sessions.sessions) {
-        auto it = saved_contexts.find(session_id);
-        if(it == saved_contexts.end()) {
-            continue;
-        }
-        auto& saved = it->second;
-        auto host_id = saved.host_path_id;
-        auto& occurrence = saved.occurrence;
-        bool orphaned = false;
-        if(host_id != no_path_id) {
-            orphaned = workspace.dep_graph.find_include_chain(host_id, session_id).empty();
-            // A pinned occurrence can vanish while other inclusions of the
-            // header survive (the chain stays non-empty) — recount it.
-            if(!orphaned && occurrence.has_value()) {
-                auto count = workspace.count_occurrences(host_id, session_id);
-                orphaned = count > 0 && *occurrence >= count;
-            }
-            // The pinned host command itself can vanish (a CDB reload
-            // changed the entry's flags): same validation didOpen applies.
-            if(!orphaned && !saved.command_hash.empty()) {
-                orphaned =
-                    !entry_has_hash(workspace.path_pool.resolve(host_id), saved.command_hash);
-            }
-        } else if(!saved.command_hash.empty()) {
-            // Own-entry pin: the pinned command must still exist in the CDB.
-            orphaned = !entry_has_hash(workspace.path_pool.resolve(session_id), saved.command_hash);
-        }
-        if(orphaned) {
-            LOG_INFO("Dropping orphaned context choice for {}: its basis no longer exists",
-                     workspace.path_pool.resolve(session_id));
-            drop_header_context(session_id);
-            SessionStore::reset_compile_state(*session, ResetDepth::Superseded);
-            saved_contexts.erase(it);
-            dropped_saved = true;
-        }
-    }
-    return dropped_saved;
-}
-
-ext::QueryContextResult ContextResolver::query_contexts(llvm::StringRef path,
-                                                        std::uint32_t path_id,
-                                                        const ext::QueryContextParams& params) {
-    auto& ws = workspace;
-    int offset_val = std::max(0, params.offset.value_or(0));
-    constexpr int page_size = 10;
-
-    ext::QueryContextResult result;
-    std::vector<ext::ContextItem> all_items;
-
-    // Contexts that would produce identical compilation results are
-    // collapsed: identical canonical flags mean an identical compile
-    // — but only for headers CONFIRMED self-contained. A header that
-    // needs includer context gets a different synthesized prefix per
-    // host, and an un-trialed header may turn out the same way, so
-    // every host stays a distinct context for both.
-    llvm::StringSet<> seen_configs;
-    bool dedup_hosts = header_mode(path, path_id) == HeaderMode::SelfContained;
-
-    auto hosts = ws.dep_graph.find_host_sources(path_id);
-    for(auto host_id: ws.rank_hosts(path_id, hosts)) {
-        auto host_path = ws.path_pool.resolve(host_id);
-        if(!ws.cdb.has_entry(host_path))
-            continue;
-        auto host_uri_opt = lsp::URI::from_file_path(std::string(host_path));
-        if(!host_uri_opt)
-            continue;
-
-        // A multi-configuration host contributes one context per
-        // CDB entry: each configuration compiles the header under
-        // different preprocessor state.
-        std::vector<std::string> host_append, host_remove;
-        ws.config.match_rules(host_path, host_append, host_remove);
-        auto cmds = ws.cdb.lookup(host_path, {.remove = host_remove, .append = host_append});
-        auto occurrences = ws.count_occurrences(host_id, path_id);
-
-        for(auto& cmd: cmds) {
-            auto hash = canonical_command_hash(cmd.to_string_argv(), cmd.resolved.directory);
-            if(dedup_hosts && !seen_configs.insert(hash).second)
-                continue;
-
-            ext::ContextItem item;
-            item.label = llvm::sys::path::filename(host_path).str();
-            if(cmds.size() > 1) {
-                auto desc = flags_label(cmd);
-                if(!desc.empty()) {
-                    item.label = std::format("{} [{}]", item.label, desc);
-                }
-                item.command_hash = hash;
-            }
-            item.description = std::string(host_path);
-            item.uri = host_uri_opt->str();
-
-            // A guard-less header can be included several times by
-            // one host — each occurrence is a distinct context.
-            if(occurrences > 1) {
-                for(std::uint32_t n = 0; n < occurrences; ++n) {
-                    auto occ_item = item;
-                    occ_item.label = std::format("{} (#{})", item.label, n + 1);
-                    occ_item.occurrence = n;
-                    all_items.push_back(std::move(occ_item));
-                }
-            } else {
-                all_items.push_back(std::move(item));
-            }
-        }
-    }
-
-    // Real entries only: lookup() would synthesize a default command
-    // even for unknown files, offering a bogus context that
-    // switchContext would then reject. Offered even when hosts
-    // exist, so a host override can be switched back to the file's
-    // own command.
-    if(ws.cdb.has_entry(path)) {
-        std::vector<std::string> rule_append, rule_remove;
-        ws.config.match_rules(path, rule_append, rule_remove);
-        auto entries = ws.cdb.lookup(path, {.remove = rule_remove, .append = rule_append});
-        auto uri_opt = lsp::URI::from_file_path(std::string(path));
-        for(std::size_t i = 0; uri_opt && i < entries.size(); ++i) {
-            auto& cmd = entries[i];
-            auto hash = canonical_command_hash(cmd.to_string_argv(), cmd.resolved.directory);
-            if(!seen_configs.insert(hash).second)
-                continue;
-
-            auto desc = flags_label(cmd);
-            ext::ContextItem item;
-            item.label = desc.empty() ? std::format("config #{}", i) : desc;
-            item.description = cmd.resolved.directory.str();
-            item.uri = uri_opt->str();
-            item.command_hash = std::move(hash);
-            all_items.push_back(std::move(item));
-        }
-    }
-
-    result.epoch = ws.context_epoch;
-    result.total = static_cast<int>(all_items.size());
-    int end = std::min(offset_val + page_size, static_cast<int>(all_items.size()));
-    for(int i = offset_val; i < end; ++i) {
-        result.contexts.push_back(std::move(all_items[i]));
-    }
-    return result;
-}
-
-ext::CurrentContextResult
-    ContextResolver::current_context(llvm::StringRef path,
-                                     const Session* session,
-                                     const ext::CurrentContextParams& params) {
-    ext::CurrentContextResult result;
-    const SavedContext* choice = session ? active_choice(session, session->path_id) : nullptr;
-    if(choice && choice->host_path_id != no_path_id) {
-        auto ctx_path = workspace.path_pool.resolve(choice->host_path_id);
-        auto ctx_uri_opt = lsp::URI::from_file_path(std::string(ctx_path));
-        if(ctx_uri_opt) {
-            ext::ContextItem item;
-            item.label = llvm::sys::path::filename(ctx_path).str();
-            if(choice->occurrence.value_or(0) > 0) {
-                item.label = std::format("{} (#{})", item.label, *choice->occurrence + 1);
-            }
-            item.description = std::string(ctx_path);
-            item.uri = ctx_uri_opt->str();
-            item.occurrence = choice->occurrence;
-            if(!choice->command_hash.empty()) {
-                item.command_hash = choice->command_hash;
-            }
-            result.context = std::move(item);
-        }
-    } else if(choice && !choice->command_hash.empty()) {
-        auto& ws = workspace;
-        ext::ContextItem item;
-        item.uri = params.uri;
-        item.command_hash = choice->command_hash;
-        item.label = std::format("config {}", choice->command_hash.substr(0, 8));
-        if(ws.cdb.has_entry(path)) {
-            std::vector<std::string> rule_append, rule_remove;
-            ws.config.match_rules(path, rule_append, rule_remove);
-            for(auto& cmd: ws.cdb.lookup(path, {.remove = rule_remove, .append = rule_append})) {
-                if(canonical_command_hash(cmd.to_string_argv(), cmd.resolved.directory) ==
-                   choice->command_hash) {
-                    auto desc = flags_label(cmd);
-                    if(!desc.empty()) {
-                        item.label = std::move(desc);
-                    }
-                    item.description = cmd.resolved.directory.str();
-                    break;
-                }
-            }
-        }
-        result.context = std::move(item);
-    }
-    return result;
-}
-
-ext::SwitchContextResult ContextResolver::switch_context(llvm::StringRef path,
-                                                         std::uint32_t path_id,
-                                                         Session* session,
-                                                         llvm::StringRef context_path,
-                                                         std::uint32_t context_path_id,
-                                                         const ext::SwitchContextParams& params) {
-    auto& ws = workspace;
-
-    ext::SwitchContextResult result;
-
-    // A choice made against an outdated listing may reference
-    // contexts that no longer exist — make the client re-query.
-    if(params.epoch.has_value() && *params.epoch != ws.context_epoch) {
-        result.stale = true;
-        return result;
-    }
-
-    if(!session) {
-        return result;
-    }
-
-    // Validate that `hash` names a real CDB entry of `entry_path`.
-    auto has_command = [&](llvm::StringRef entry_path, llvm::StringRef hash) {
-        if(!ws.cdb.has_entry(entry_path)) {
-            return false;
-        }
-        std::vector<std::string> rule_append, rule_remove;
-        ws.config.match_rules(entry_path, rule_append, rule_remove);
-        for(auto& cmd: ws.cdb.lookup(entry_path, {.remove = rule_remove, .append = rule_append})) {
-            if(canonical_command_hash(cmd.to_string_argv(), cmd.resolved.directory) == hash) {
-                return true;
-            }
-        }
-        return false;
-    };
-
-    SavedContext saved;
-    if(context_path_id == path_id && params.command_hash.has_value()) {
-        // Pin one of the file's own CDB entries.
-        if(!has_command(path, *params.command_hash)) {
-            return result;
-        }
-        saved.command_hash = *params.command_hash;
-    } else {
-        // Pin a host source for a header: it must have a real CDB
-        // entry, actually (transitively) include this header, and —
-        // for multi-configuration hosts — own the pinned entry.
-        if(!ws.cdb.has_entry(context_path)) {
-            return result;
-        }
-        if(ws.dep_graph.find_include_chain(context_path_id, path_id).empty()) {
-            return result;
-        }
-        if(params.command_hash.has_value() && !has_command(context_path, *params.command_hash)) {
-            return result;
-        }
-        if(params.occurrence.has_value() && *params.occurrence > 0) {
-            auto count = ws.count_occurrences(context_path_id, path_id);
-            if(count > 0 && *params.occurrence >= count) {
-                return result;
-            }
-        }
-        saved.host_path_id = context_path_id;
-        saved.occurrence = params.occurrence;
-        saved.command_hash = params.command_hash.value_or("");
-    }
-
-    drop_header_context(path_id);
-    // The new context is a different compilation identity: supersede any
-    // in-flight compile and drop the state earned under the old one. It
-    // also needs its own self-containment trial — a different host can
-    // change the macro environment.
-    SessionStore::reset_compile_state(*session, ResetDepth::Superseded);
-    forget_self_contained(path_id);
-
-    // The table entry is the active choice; persist it across sessions.
-    saved_contexts[path_id] = std::move(saved);
-    ws.save_cache(*this);
-
-    result.success = true;
-    return result;
 }
 
 }  // namespace clice
