+++ src/server/service/feature_router.h
@@ -4,10 +4,10 @@
 #include <vector>
 
 #include "feature/feature.h"
+#include "sched/workspace.h"
 #include "server/service/query.h"
 #include "server/state/session.h"
 #include "server/state/session_store.h"
-#include "server/state/workspace.h"
 
 #include "kota/async/async.h"
 #include "kota/codec/json/json.h"
