+++ src/server/transport/uri.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <string>
+
+namespace clice {
+
+/// Convert a file:// URI to a local file path.
+std::string uri_to_path(const std::string& uri);
+
+}  // namespace clice
