+++ tests/unit/server/compiler_tests.cpp
@@ -4,8 +4,8 @@
 
 #include "test/temp_dir.h"
 #include "test/test.h"
+#include "sched/context.h"
 #include "server/compiler/compiler.h"
-#include "server/compiler/context_resolver.h"
 #include "server/worker_test_helpers.h"
 #include "support/anomaly.h"
 #include "support/cache_store.h"
