+++ tests/unit/server/quarantine_tests.cpp
@@ -1,6 +1,7 @@
 #include <chrono>
 
 #include "test/test.h"
+#include "sched/crash_budget.h"
 #include "server/state/quarantine.h"
 
 namespace clice::testing {
