+++ tests/unit/server/query_freshness_tests.cpp
@@ -7,7 +7,7 @@
 #include "test/tester.h"
 #include "index/shard.h"
 #include "index/tu_index.h"
-#include "server/compiler/context_resolver.h"
+#include "sched/context.h"
 #include "server/compiler/indexer.h"
 #include "server/service/query.h"
 #include "server/state/session_store.h"
